payment logic

+++ src/main/java/com/group02/openevent/repository/ICompetitionEventRepo.java
@@ -3,7 +3,7 @@ package com.group02.openevent.repository;
 import com.group02.openevent.model.event.CompetitionEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICompetitionEventRepo extends JpaRepository<CompetitionEvent, Integer> {
+public interface ICompetitionEventRepo extends JpaRepository<CompetitionEvent, Long> {
 
 }
