+++ src/main/java/com/group02/openevent/repository/IWorkshopEventRepo.java
@@ -3,7 +3,7 @@ package com.group02.openevent.repository;
 import com.group02.openevent.model.event.WorkshopEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IWorkshopEventRepo extends JpaRepository<WorkshopEvent, Integer> {
+public interface IWorkshopEventRepo extends JpaRepository<WorkshopEvent, Long> {
 
 }
